fix(comparisons): detect mismatched Flint implementation statuses
