+++ packages/comparisons/package.json
@@ -29,7 +29,16 @@
 		"prettier": "3.8.1"
 	},
 	"devDependencies": {
+		"@flint.fyi/astro": "workspace:^",
+		"@flint.fyi/browser": "workspace:^",
+		"@flint.fyi/jsx": "workspace:^",
+		"@flint.fyi/node": "workspace:^",
+		"@flint.fyi/performance": "workspace:^",
+		"@flint.fyi/plugin-flint": "workspace:^",
+		"@flint.fyi/spelling": "workspace:^",
+		"@flint.fyi/vitest": "workspace:^",
 		"eslint": "catalog:dev",
+		"flint": "workspace:^",
 		"tsdown": "catalog:dev",
 		"vitest": "catalog:dev"
 	},
