+++ packages/comparisons/src/data.test.ts
@@ -1,27 +1,13 @@
 import { builtinRules } from "eslint/use-at-your-own-risk";
 import { describe, expect, it } from "vitest";
 
+import { getRuleForPluginSafe } from "./getRuleForPlugin.ts";
 import { comparisons, getComparisonId } from "./index.ts";
 import { groupByLinterAndPlugin } from "./test-util.ts";
 
 const groupedData = groupByLinterAndPlugin(comparisons);
 
 describe("data.json", () => {
-	it("does not include any duplicate Flint rules", () => {
-		const seenIds = new Set<string>();
-
-		for (const comparison of comparisons) {
-			const id = getComparisonId(
-				comparison.flint.plugin,
-				comparison.flint.name,
-			);
-
-			expect(seenIds).not.toContain(id);
-
-			seenIds.add(id);
-		}
-	});
-
 	describe("Comparison with ESLint", () => {
 		it("includes all builtin rules", () => {
 			const builtinESLintRuleNames = new Set<string>(
@@ -43,6 +29,51 @@ describe("data.json", () => {
 			expect(builtinESLintRuleNamesCoveredByFlint).toEqual(
 				builtinESLintRuleNames,
 			);
+		});
+	});
+
+	describe("Comparison with Flint", () => {
+		it("does not include any duplicate Flint rules", () => {
+			const seenIds = new Set<string>();
+
+			for (const comparison of comparisons) {
+				const id = getComparisonId(
+					comparison.flint.plugin,
+					comparison.flint.name,
+				);
+
+				expect(seenIds).not.toContain(id);
+
+				seenIds.add(id);
+			}
+		});
+
+		describe("data matching", () => {
+			const pairs = comparisons.map((comparison) => ({
+				comparison,
+				rule: getRuleForPluginSafe(
+					comparison.flint.plugin,
+					comparison.flint.name,
+				),
+			}));
+
+			it("does not mark rules as implemented that do not exist yet", () => {
+				expect(
+					pairs.filter(
+						({ comparison, rule }) =>
+							!rule && comparison.flint.status === "implemented",
+					),
+				).toEqual([]);
+			});
+
+			it("does not miss rules as implemented that exist", () => {
+				expect(
+					pairs.filter(
+						({ comparison, rule }) =>
+							rule && comparison.flint.status !== "implemented",
+					),
+				).toEqual([]);
+			});
 		});
 	});
 });
