+++ packages/comparisons/src/getRuleForPlugin.ts
