+++ packages/site/src/components/RuleSummary.astro
@@ -1,6 +1,6 @@
 ---
 import markdownIt from "markdown-it";
-import { getRuleForPlugin } from "./getRuleForPlugin";
+import { getRuleForPlugin } from "@flint.fyi/comparisons";
 
 interface Props {
 	plugin: string;
